refactor component folders
