+++ src/components/Header/Create.jsx
@@ -1,8 +1,8 @@
 import { useState } from "react";
-import { getUserByUsername, getUsers, addUser } from "../utils/Api";
+import { getUserByUsername, getUsers, addUser } from "../../utils/Api";
 import { useNavigate, Link } from "react-router-dom";
-import { useLoggedUser, useHasCreatedState } from "../contexts/AllContexts";
-import FooterCredits from "./FooterCredits.jsx";
+import { useLoggedUser, useHasCreatedState } from "../../contexts/AllContexts";
+import FooterCredits from "../UI/FooterCredits.jsx";
 import { createClient } from "@supabase/supabase-js";
 
 const supabase = createClient(
